refactor: Board is now Game
