+++ src/game/mod.rs
@@ -1,2 +1,1 @@
-pub mod board;
 pub mod game;
