+++ src/handler.rs
@@ -1,7 +1,7 @@
 use crate::{
     app::{App, AppResult},
     constants::Pages,
-    game::board::Coord,
+    game::game::Coord,
 };
 use ratatui::crossterm::event::{
     KeyCode, KeyEvent, KeyEventKind, KeyModifiers, MouseButton, MouseEvent, MouseEventKind,
@@ -13,14 +13,14 @@ pub fn handle_key_events(key_event: KeyEvent, app: &mut App) -> AppResult<()> {
         // crossterm on Windows sends Release and Repeat events as well, which we ignore.
         return Ok(());
     }
-    if app.game.board.mouse_used {
-        app.game.board.mouse_used = false;
-        if app.game.board.selected_coordinates != Coord::undefined() {
-            app.game.board.cursor_coordinates = app.game.board.selected_coordinates;
-            app.game.board.selected_coordinates = Coord::undefined();
+    if app.game.mouse_used {
+        app.game.mouse_used = false;
+        if app.game.selected_coordinates != Coord::undefined() {
+            app.game.cursor_coordinates = app.game.selected_coordinates;
+            app.game.selected_coordinates = Coord::undefined();
         } else {
-            app.game.board.cursor_coordinates.col = 4;
-            app.game.board.cursor_coordinates.row = 4;
+            app.game.cursor_coordinates.col = 4;
+            app.game.cursor_coordinates.row = 4;
         }
     }
 
@@ -41,7 +41,7 @@ pub fn handle_key_events(key_event: KeyEvent, app: &mut App) -> AppResult<()> {
             if app.current_page == Pages::Bot && app.selected_color.is_none() {
                 app.menu_cursor_right(2);
             } else {
-                app.game.board.cursor_right();
+                app.game.cursor_right();
             }
         }
         KeyCode::Left | KeyCode::Char('h') => {
@@ -49,21 +49,21 @@ pub fn handle_key_events(key_event: KeyEvent, app: &mut App) -> AppResult<()> {
             if app.current_page == Pages::Bot && app.selected_color.is_none() {
                 app.menu_cursor_left(2);
             } else {
-                app.game.board.cursor_left();
+                app.game.cursor_left();
             }
         }
         KeyCode::Up | KeyCode::Char('k') => {
             if app.current_page == Pages::Home {
                 app.menu_cursor_up(Pages::variant_count() as u8);
             } else {
-                app.game.board.cursor_up();
+                app.game.cursor_up();
             }
         }
         KeyCode::Down | KeyCode::Char('j') => {
             if app.current_page == Pages::Home {
                 app.menu_cursor_down(Pages::variant_count() as u8);
             } else {
-                app.game.board.cursor_down();
+                app.game.cursor_down();
             }
         }
         KeyCode::Char(' ') | KeyCode::Enter => {
@@ -72,7 +72,7 @@ pub fn handle_key_events(key_event: KeyEvent, app: &mut App) -> AppResult<()> {
             } else if app.current_page == Pages::Home {
                 app.menu_select();
             } else {
-                app.game.board.select_cell();
+                app.game.select_cell();
             }
         }
         KeyCode::Char('?') => {
@@ -91,7 +91,7 @@ pub fn handle_key_events(key_event: KeyEvent, app: &mut App) -> AppResult<()> {
                 app.show_color_popup = false;
                 app.menu_cursor = 0;
             }
-            app.game.board.unselect_cell();
+            app.game.unselect_cell();
         }
         KeyCode::Char('b') => {
             app.go_to_home();
@@ -108,7 +108,7 @@ pub fn handle_mouse_events(mouse_event: MouseEvent, app: &mut App) -> AppResult<
         return Ok(());
     }
     if mouse_event.kind == MouseEventKind::Down(MouseButton::Left) {
-        if app.game.board.is_checkmate || app.game.board.is_draw {
+        if app.game.is_checkmate || app.game.is_draw {
             return Ok(());
         }
 
@@ -118,26 +118,26 @@ pub fn handle_mouse_events(mouse_event: MouseEvent, app: &mut App) -> AppResult<
 
         // If there is a promotion to be done the top_x, top_y, width and height
         // values are updated accordingly
-        if app.game.board.is_promotion {
-            let x = (mouse_event.column - app.game.board.top_x) / app.game.board.width;
-            let y = (mouse_event.row - app.game.board.top_y) / app.game.board.height;
+        if app.game.is_promotion {
+            let x = (mouse_event.column - app.game.top_x) / app.game.width;
+            let y = (mouse_event.row - app.game.top_y) / app.game.height;
             if x > 3 || y > 0 {
                 return Ok(());
             }
-            app.game.board.promotion_cursor = x as i8;
-            app.game.board.promote_piece();
+            app.game.promotion_cursor = x as i8;
+            app.game.promote_piece();
         }
-        if mouse_event.column < app.game.board.top_x || mouse_event.row < app.game.board.top_y {
+        if mouse_event.column < app.game.top_x || mouse_event.row < app.game.top_y {
             return Ok(());
         }
-        let x = (mouse_event.column - app.game.board.top_x) / app.game.board.width;
-        let y = (mouse_event.row - app.game.board.top_y) / app.game.board.height;
+        let x = (mouse_event.column - app.game.top_x) / app.game.width;
+        let y = (mouse_event.row - app.game.top_y) / app.game.height;
         if x > 7 || y > 7 {
             return Ok(());
         }
-        app.game.board.mouse_used = true;
+        app.game.mouse_used = true;
         let coords: Coord = Coord::new(y as u8, x as u8);
-        app.game.board.move_selected_piece_cursor_mouse(coords);
+        app.game.move_selected_piece_cursor_mouse(coords);
     }
     Ok(())
 }
