+++ tests/pieces/knight.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-    use chess_tui::game::board::{Board, Coord};
+    use chess_tui::game::game::{Coord, Game};
     use chess_tui::pieces::knight::Knight;
     use chess_tui::pieces::{PieceColor, PieceType, Position};
     use chess_tui::utils::is_getting_checked;
@@ -26,7 +26,7 @@ mod tests {
             [None, None, None, None, None, None, None, None],
             [None, None, None, None, None, None, None, None],
         ];
-        let mut board = Board::default();
+        let mut board = Game::default();
         board.set_board(custom_board);
 
         let mut right_positions = vec![
@@ -92,7 +92,7 @@ mod tests {
                 Some((PieceType::Knight, PieceColor::White)),
             ],
         ];
-        let mut board = Board::default();
+        let mut board = Game::default();
         board.set_board(custom_board);
 
         let mut right_positions = vec![Coord::new(6, 5)];
@@ -149,7 +149,7 @@ mod tests {
                 Some((PieceType::Knight, PieceColor::Black)),
             ],
         ];
-        let mut board = Board::new(custom_board, PieceColor::White, vec![]);
+        let mut board = Game::new(custom_board, PieceColor::White, vec![]);
         board.set_board(custom_board);
 
         let is_king_checked =
@@ -209,7 +209,7 @@ mod tests {
                 Some((PieceType::Knight, PieceColor::Black)),
             ],
         ];
-        let mut board = Board::new(custom_board, PieceColor::White, vec![]);
+        let mut board = Game::new(custom_board, PieceColor::White, vec![]);
         board.set_board(custom_board);
 
         let is_king_checked =
@@ -268,7 +268,7 @@ mod tests {
             [None, None, None, None, None, None, None, None],
             [None, None, None, None, None, None, None, None],
         ];
-        let mut board = Board::new(custom_board, PieceColor::Black, vec![]);
+        let mut board = Game::new(custom_board, PieceColor::Black, vec![]);
         board.set_board(custom_board);
 
         let is_king_checked =
